Egl display works.
Add some log, which will be removed later.

+++ shmem_ivshm_ivshmem.c
@@ -69,7 +69,7 @@ static int shmem_open(const char *devpath, struct shmem_info *info, int evt_fds[
 	if (ioctl(iregion_fd, IVSHM_GET_MMIO_SZ, &info->mem_size) < 0)
 		error(1, errno, "failed to get ivshm mmio size");
 
-	printf("%s, mem_size: 0x%lx___\r\n", __func__, info->mem_size);
+	printf("%s, mem_size: 0x%lx___\r\n", __func__, (unsigned long)info->mem_size);
 	info->mem_base = mmap(NULL, info->mem_size, PROT_READ | PROT_WRITE, MAP_SHARED, iregion_fd, 0);
 	if (info->mem_base == MAP_FAILED)
 		error(1, errno, "mmap of shared memory failed");
